+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 > All notable changes to this project will be documented in this file.
 
+
+## v1.1.4
+
+Released: January 31, 2021
+
+### Features v1.1.4
+
+- **BsMaskLoader:** add properties `overlay-color`, `overlay-opacity`, `spinner-color`,
+  `spinner-thickness`, and `spinner-type`
+
+### Bug Fixes & Improvement v1.1.4
+
+- **BsAppbarItems:** remove `font-size` scss styles
+- **BsMenu:** improve keyboard navigation using `arrow-down` and `arrow-up`
+- **BsPopover:** fix `space` property did not apply correctly when `placement` property sets to 
+  `top`, `top-left` or `top-right`
+- Refactor `popover` transition animation to `scale`  
+
+
 ## v1.1.3
 
 Released: January 5, 2021
@@ -9,9 +28,10 @@ Released: January 5, 2021
 ### Bug Fixes & Improvement v1.1.3
 
 - **BsGrid:** fix side effect on methods `sort`, `setPageSize`, `_fetchData`, and update scss styles 
-- **BsSearchField:** remove property `can-close` and fix `BsPopover` show/hide procedure when `search-options` is enabled
+- **BsSearchField:** remove property `can-close` and fix `BsPopover` show/hide procedure when 
+  `search-options` is enabled
 - **BsTreegridCell:** fix warning when icon is null
-- **BsCheckboxGroup, BsRadioGroup:** fix multi-column items when `columns` property is set to `1` 
+- **BsCheckboxGroup, BsRadioGroup:** fix multi-column items when `columns` property value is `1` 
 - **BsButtonToggleField:** fix non-persistent help text and update scss styles
 - **AbstractStore:** add methods `setPageSize` and `setSorters`
 - **BsStore:** fix method `append` didn't send data to remote service
@@ -27,7 +47,8 @@ Released: December 21, 2020
 - **Touch** directive did not expose correctly
 - **BsCard:** add DOM event listener
 - **BsNotification:** change default variant and update css styles
-- **BsTextField, BsTextArea, BsRadioGroup, BsDateTimeField, BsCombobox, BsCheckboxGroup:** bug fix when form reset
+- **BsTextField, BsTextArea, BsRadioGroup, BsDateTimeField, BsCombobox, BsCheckboxGroup:** bug fix 
+  when form reset
 
 
 ## v1.1.0
@@ -104,7 +125,7 @@ Released: Mei 20, 2020
 
 ### Bug Fixes v1.0.3
 
-- **BsSpacer:** can't fill available space
+- **BsSpacer:** fix didn't fill available space
 
 
 ## v1.0.2
@@ -113,6 +134,6 @@ Released: Mei 20, 2020
 
 ### Bug Fixes v1.0.2
 
-- **BsGrid:** fix localPaging, localSorting, and localFiltering
+- **BsGrid:** fix local paging, local sorting, and local filtering
 - **BsPagination:** fix combobox paging
 - **BsArrayStore, BsStore:** enhance LocalFilter to match BsGrid filterable spec